+++ cqe-challenges/KPRCA_00048/cb_2/lib/stdint.h
@@ -11,8 +11,8 @@ typedef unsigned short int uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long long int uint64_t;
 
-typedef int intptr_t;
-typedef unsigned int uintptr_t;
+typedef long int intptr_t;
+typedef unsigned long int uintptr_t;
 
 #  define __INT64_C(c)  c ## LL
 #  define __UINT64_C(c) c ## ULL
